# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  description: |
    API for WASAPhoto social Project.
    Each user will be presented with a stream of photos (images) in reverse chronological order, with
    information about when each photo was uploaded (date and time) and how many likes and comments
    it has. The stream is composed by photos from “following” (other users that the user follows). Users
    can place (and later remove) a “like” to photos from other users. Also, users can add comments to any
    image (even those uploaded by themself). Only authors can remove their comments.
    Users can ban other users. If user Alice bans user Eve, Eve won’t be able to see any information about
    Alice. Alice can decide to remove the ban at any moment.
    Users will have their profiles. The personal profile page for the user shows: the user’s photos (in reverse
    chronological order), how many photos have been uploaded, and the user’s followers and following.
    Users can change their usernames, upload photos, remove photos, and follow/unfollow other users.
    Removal of an image will also remove likes and comments.
    A user can search other user profiles via username.
    A user can log in just by specifying the username.
  version: 1.0.0
servers: 
  - url: 'http://localhost:3000'
...
paths:
...
  /users/{userId}/photos:
    parameters:
      - $ref: '#/components/parameters/userId'
    post:
      security:
      - bearerAuth : []
      tags: ["photos"]
      description: allows to upload a new photo and return the id of the photos
      summary: Upload a new photo
      operationId: uploadPhoto
      requestBody:
        required: true
        description: new photo
        content:
          multipart/form-data:
            schema:
              type: object
              description: photo to upload
              properties:
                photo:
                  description: The photo to upload
                  type: string
                  format: binary
                  minLength: 1
                  maxLength: 64
      responses:
        "201":
          description: photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
              examples:
                example1:
                  value:
                    photosId: "abc123"
                    url: "https://example.com/Users/{userId}/photos/{photosId}"
                    timestamp: "2024-05-16T17:24:49.408Z"
                    likes: ["1","32"]
                    comments: [""]
        "401":
          $ref: '#/components/responses/UnauthorizedError'
  /users/{userId}/photos/{photosId}:
    parameters:
      - $ref: '#/components/parameters/userId'    
      - $ref: '#/components/parameters/photosId'
    delete:
      security:
      - bearerAuth : []
      tags: ["photos"]
      summary: Delete a photo
      description: Deletes a photo belonging to the authenticated user.
      operationId: deletePhoto
      responses:
        '204':
          description: |
            Photo and his likes and comments deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    get:
      security:
      - bearerAuth : []
      tags: ["photos"]
      description: allows to get a photo
      summary: Get a photo
      operationId: getPhoto
      responses:
        "200":
          description: successfully return photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
              examples:
                example1:
                  value:
                    photosId: "abc123"
                    url: "https://example.com/Users/{userId}/photos/{photosId}"
                    timestamp: "2024-05-16T17:24:49.408Z"
                    likes: ["1","32"]
                    comments: [""]
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/UserNotFound'

#-------Photo likes-------#
  /users/{userId}/photos/{photosId}/likes:
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/photosId'
    post:
      security:
      - bearerAuth : []
      tags: ["likes"]
      description: allows to like photos
      summary: like photos
      operationId: likePhoto
      responses:
        "201":
          $ref: "#/components/responses/LikePhoto"
        "401":
          $ref: '#/components/responses/UnauthorizedError'
  /users/{userId}/photos/{photosId}/likes/{likeId}:
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/photosId'
      - $ref: '#/components/parameters/likeId'
    delete:
      security:
      - bearerAuth : []
      tags: ["likes"]
      description: allows to unlike Photo
      summary: unlike photos
      operationId: unlikePhoto
      responses:
        "204":
          description: photo like removed
        "401":
          $ref: '#/components/responses/UnauthorizedError'
          
#-------Photo comments-------#
  /users/{userId}/photos/{photosId}/comments:
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/photosId'
    post:
      security:
      - bearerAuth : []
      tags: ["comments"]
      description: allows to comment photos
      summary: comment photos
      operationId: commentPhoto
      responses:
        "201":
          description: photo commented successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
              examples:
                example1:
                  value:
                    commentId: "123"
                    text: "ciao come stai?"
        "401":
          $ref: '#/components/responses/UnauthorizedError'
    get:
      security:
      - bearerAuth : []
      tags: ["comments"]
      description: get comments on a photo
      summary: retrieve comments for a photo
      operationId: GetPhotoComments
      responses:
        "200":
          description: successfully retrieve comments
          content:
            application/json:
              schema:
                description: list of comments
                type: object
                properties:
                  comments:
                    description: list of comments
                    type: array
                    minItems: 0
                    maxItems: 1000
                    items: 
                      $ref: "#/components/schemas/Comment"
              examples:
                example:
                  value:
                    comments:
                      - commentId: "123"
                        userId: "1"
                        text: "Nice photo"
                      - commentId: "12"
                        userId: "32"
                        text: "You are great"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404": 
          $ref: "#/components/responses/UserNotFound"
  /users/{userId}/photos/{photosId}/comments/{commentId}:
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/photosId'
      - $ref: '#/components/parameters/commentId'
    delete:
      security:
      - bearerAuth : []
      tags: ["comments"]
      description: allows to delete comment on photos
      summary: comments on photos
      operationId: uncommentPhoto
      responses:
        "204":
          description: comment deleted successfully
        "401":
          $ref: '#/components/responses/UnauthorizedError'


#-------follows e followers-------#
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: string
      description: User ID used as bearer token
  responses:
    LikePhoto:
      description: Like a photo
      content:
        application/json:
          schema:
            description: Like a photo
            type: object
            properties:
              userId:
                description: user to like photo
                type: string
                example: "123"
              userPhotoId:
                description: user that posted the photo
                type: string
                example: "1"
              photoId:
                description: Photo unique identifier
                type: string
                example: "1234"
              likeId:
                description: like unique identifier
                type: string
                example: "134"
...
    UnauthorizedError:
      description: Access identifier is missing or invalid
      content:
        application/json:
          schema:
            description: Unauthorized error
            type: object
            properties:
              message:
                description: Error message
                type: string
                example: Unauthorized
              code:
                type: integer
                example: 401
                description: HTTP status code
...
    UserNotFound:
      description: A user with the specified ID was not found.
      content:
        application/json:
          schema:
            description: User not found error
            type: object
            properties:
              message:
                description: Error message
                type: string
                example: User not found
              code:
                type: integer
                example: 404
                description: HTTP status code
...
  parameters:
...
    userId:
      name: userId
      in: path
      required: true
      description: ID of the user
      schema:
        description: ID of the user
        type: string    
        pattern: '^.*?$'
        minLength: 1
        maxLength: 20
    commentId:
      name: commentId
      in: path
      required: true
      description: ID of the comment to delete
      schema:
        description: ID of the comment to delete
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 20
    likeId:
      name: likeId
      in: path
      required: true
      description: ID of the like to delete
      schema:
        description: ID of the like to delete
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 20
...
    photosId:
      name: photosId
      in: path
      description: ID of the photo
      required: true
      schema:
        description: ID of the photo
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 20
  schemas:
...
    Photo:
      description: Photo details
      type: object
      properties:
        photoId:
          type: string
          description: The unique identifier of the photo
        url:
          type: string
          format: url
          description: The URL where the photo is stored in the database
        timestamp:
          type: string
          format: date-time
          description: The date and time when the photo was uploaded
        likes:
          type: array
          minItems: 0
          maxItems: 100
          items:
            type: string
            description: IDs of the user who liked the photo
          description: user who liked the photo
        comments:
          type: array
          minItems: 0
          maxItems: 100
          items:
            type: string
            description: text of the comment
          description: list of comments
    Comment:
      description: Comment details
      type: object
      properties:
        userId:
          type: string
          description: The unique identifier of the user who commented
        photosId:
          type: string
          description: The unique identifier of the photo
        commentId:
          type: string
          description: The unique identifier of the comment
        text:
          type: string
          description: text of the comment
...
security:
  - bearerAuth: []